Implements a new "control socket" feature.
There was already a unix socket acting as a lock,
this adds the possibility to listen to it.

Document uploads now use http, previously it used a socket,
but using http makes more sense anyway (might be a bit slower).

Also implements new log file cleaner (for on the fly compressed log files).